Crate sel4_rust_doc

WORK IN PROGRESS

This library is not even close to being suitable as a reference. Please refer to documentation provided by the seL4 project.

Attribution

The vast majority of the doc-comment content in this library is either copied from the seL4 manual verbatim, or only slightly changed in words but not meaning. See also acknowledgements

Introduction

This library is based on on the seL4 manual, and attribution for the text content, design, etc. goes to this project as a whole. I can only take credit for transcribing my percieved intent of the manual into this form.

The seL4 microkernel is an operating-system kernel designed to be a secure, safe, and reliable foundation for systems in a wide variety of application domains. As a microker- nel, it provides a small number of services to applications, such as abstractions to create and manage virtual address spaces, threads, and inter-process communication (IPC). The small number of services provided by seL4 directly translates to a small imple- mentation of approximately 8700 lines of C code. This has allowed the ARMv6 version of the kernel to be formally proven in the Isabelle/HOL theorem prover to adhere to its formal specification [Boy09,CKS08,DEK+06,EKE08,KEH+09,TKN07,WKS+09], which in turn enabled proofs of the kernel’s enforcement of integrity [SWG+11] and confidentiality [MMB+13]. The kernel’s small size was also instrumental in performing a complete and sound analysis of worst-case execution time [BSC+11,BSH12]. This library documents the seL4 kernel’s API from a user’s point of view. The document starts by giving a brief overview of the seL4 microkernel design, followed by a reference of the high-level API exposed by the seL4 kernel to userspace.

The manual authors tried to ensure their work accurately reflects the behaviour of the seL4 kernel, it is by no means a formal specification of the kernel. When the precise behaviour of the kernel under a particular circumstance needs to be known, users should refer to the seL4 abstract specification, which gives a formal description of the seL4 kernel. This can be found in the seL4 projects github account under the l4v repo.

Rust library author comment

This project intends to be a good-faith effort of replicating what the manual aims to communicate, without changing its meaning or detracting from its qualities. It is currently an independent project.

The intent is to leverage Rusts expressive type system Idiomatic Rust equivilant might follow these examples

C “0 on success, error data arrangement depending on error” becomes Result<(), SeL4Error>, which is further aliased to SeL4Result<()>

When an operation has an “optional extension”, and opting into this extension is achieved by calling a different function with an added argument, these are can be combined into a single function with an Option<_> argument to differentiate.

Use traits

Where the api is architecture defined, his library is limited to the aarch64 platform for the time being.

Modules

acknowledgements
Content of manuals Acknowledgements.

bibliography
Direct copy-paste of the raw text from the manual bibliography section

kernel_api
A small set of services and objects