Struct sel4_rust_doc::kernel_api::object_interfaces::capability_space::CapSpace

pub struct CapSpace {
    pub root_cnode: CapNode,
}

A root CapNode, allowing a ThreadControlBlock to manage its capabilities

Fields
root_cnode: CapNode
Implementations

impl CapSpace

Rust equivilent of seL4_CNode_${Func} functions

pub fn mint(
&mut self,
src_slot: Slot,
dest_root: Option<&mut CapSpace>,
dest_slot: Slot,
rights: CapRights,
badge: Option<Badge>
) -> SeL4Result<()>

Copy a capability, setting its rights in the process

if badge is None, then this is equivilant to seL4_CNode_Copy

if dest_root is Some, then the copy ends up in the dest cap_space.

if dest_root is None, then dest_slot must not equal src_slot, otherwise the clash will result in an error.

pub fn move_(
&mut self,
src_slot: Slot,
dest_root: Option<&mut CapNode>,
dest_slot: Slot,
mutation: Option<Badge>
) -> SeL4Result<()>

Moves a capability from an occupied slot to an empty slot

If mutation is a value of Some(_), then it is the equivilant of seL4_CNode_Mutate

same dest_root and slot rules apply as mint

pub fn rotate(
dest_slot: Slot,
pivot_root: Option<&mut CapNode>,
pivot_slot: Slot,
source_root: Option<&mut CapNode>,
source_slot: Slot
) -> SeL4Result<()>

Two moves in a single, atomic operation

The pivot slot must be distinct from the source and destination The destination slot must be empty, unless it’s the same as the source, in which case, its content will be swapped with the pivot slot

analagous to the following, only done atomically

// src != dest
pivot_cspace.move_(pivot, dest_cspace, dest, None);
src_cspace.move_(src, pivot_cspace, pivot, None);
// src == dst
src_cspace.move_(src, tmp_cspace, tmp, None);
// or pivot_cspace.move_(pivot, src_cspace, src, None), as src == dest
pivot_cspace.move_(pivot, dest_cspace, dest);
temp_cspace.move_(temp, pivot_cspace, pivot, None);

pub fn delete(&mut self, slot: Slot) -> SeL4Result<()>

Removes the capability

pub fn revoke(&mut self, slot: Slot) -> SeL4Result<()>

Equivilent to CapSpace::delete on each capability derived from slot

Refer to UntypedMemory documentation for further details on capability derivation.

pub fn save_caller(root_capnode: &mut CapNode, slot: Slot) -> SeL4Result<()>

Save the kernel generated reply capability from the most recent time the thread was called, placing it into this CapSpace so it can be used later

pub fn cancel_badged_sends(&mut self, index: usize) -> SeL4Result<()>

Allows the reuse of badges by an authority.

Badged Endpoints only -> anything else, will have no effect

The badged endpoint being looked up has its list of outstanding send operations with a matching badge