Struct sel4_rust_doc::kernel_api::object_interfaces::endpoints::EndPoint

pub struct EndPoint;

Attaches to a ThreadControlBlock to become a Send/Recieve “port” for IPCs

IPC is synchronous: A thread sending/recieving on an endpoint will bloock untill delivery.

This implies delivery only happens when both sender & reciever rendezvous at the endpoint.

An end-point can be restricted to be send-only/recieve only. An end point can have the grant right: allows sending of capabilities.

Implementations

impl EndPoint

pub fn load_cap(cap: CapPtr)

adds a capability to be included in the next message send Requires Grand permissions on this endpoint

Without Grant permissions, only raw msg sent (no cap x-fer)