Module sel4_rust_doc::kernel_api::object_interfaces::reply

Reply is a special, “free” capability that is saved to a single, dedicated slot in a thread after it recieves a message. reply is equivilant to CapSpace::save_caller followed by a SeL4Send::send

Structs

Reply