Module sel4_rust_doc::kernel_api::object_interfaces::thread_control_block

Representing execution context, and an applications processor usage.

Is assinged a CapSpace and VSpace

May also be assigned an IPCBuffer, without which, a thread is limited to the architecture defined message registers.

Thread Creation

A newly created thread is initially inactive. Before activating, it requires to be assigned to it a CapSpace and VSpace (using the set_space(TODO) or configure methods) as well as setting the initial stack pointer and instruction pointer (set using the write_registers method (TODO))

Thread Activation

With the spaces and registers set-up, the TCB can be activated by either setting the resume_target parameter in write_registers to true, or by calling the resume method.

Multi-core machines

By default, the thread runs on the same CPU which originally created the TCB. It can be migrated to other CPUs using the sen_affinity method

Thread Deactivation

A thread can be suspended using the suspend method. A suspended thread can then later be resumed

A suspended state can be retrieved with the read_registers and copy_registers method

A TCB can be reconfigured and reused, or left suspended indefinately if not needed. When the last capability to a TCB is deleted, the thread will automatically suspend.

Scheduling

TODO

Exceptions

TODO

Exceptions

TODO

Message Layout of the Read-/Write-Registers methods

TODO

Faults

TODO

Structs

ThreadControlBlock
Representing an execution context. Can be scheduled, blocked, unblocked, etc.

UserContext