Trait sel4_rust_doc::kernel_api::object_interfaces::vspace::aarch64::CacheControl

pub trait CacheControl {
    fn clean_data(&mut self, start_offset: usize, end_offset_excl: usize);

    fn clean_invalidate_data(
        &mut self, 
        start_offset: usize, 
        end_offset_excl: usize
    );

    fn invalidate_data(&mut self, start_offset: usize, end_offset_excl: usize);

    fn unify_instructions(&mut self, start: usize, end: usize);
}

Required methods

fn clean_data(&mut self, start_offset: usize, end_offset_excl: usize)

cleans the data cache range given within this page out to RAM

fn clean_invalidate_data(&mut self, start_offset: usize, end_offset_excl: usize)

cleans and invalidates the cache range given within this page and flushes data out to RAM

fn invalidate_data(&mut self, start_offset: usize, end_offset_excl: usize)

Invalidates the cache range within the page. The start and end should be alligned to a cache line boundary where possible If they are not aligned, an additional clean is done to the outer cache lines in full also

fn unify_instructions(&mut self, start: usize, end: usize)

Loading content...