Module sel4_rust_doc::kernel_api::syscalls

Raw kernel interface. When invoked with a capability, allows for IPC communication and use of kernel objects.

Threads send messages by invoking capabilities within their capability space. When an EndPoint, Notification or Reply capability is invoked in this way, the message will be transferred through the kernel to another thread. When capabilities to other kernel objects are invoked, the message will be interpreted as a method invocation in a manner specific to the type of kernel object. For example, a thread of execution can be suspended by invoking a ThreadControlBlock (TCB) capability with the correctly formatted message.

Much how libc can simplify usage of a POSIX api by acting as a layer between the user and the system api itself, libsel4 acts as a layer between the user and these syscalls. For example, configuring a TCB requires the setting of 3 specific capabilities and 4 variable values in the message::Info. This is handled by libsel4 seL4_TCB_Configure.

With the exception of yield, there are only two syscalls: send and recv, with the others being only slight variations/combinations of these two.

Traits

SeL4Recv
Syscalls that recieve on a capability

SeL4Send
Syscalls that send on a capability.

SeL4SendRecv
Allows send and recieve in a single call, thus improving performance.

Functions

reply
Uses reply cap generated by call() to respond to it.

Type Definitions

SeL4Result