Trait sel4_rust_doc::kernel_api::syscalls::SeL4Recv

pub trait SeL4Recv: Capability {
    fn recv(&self, get_badge: Option<&mut Badge>) -> Info;

    fn nb_recv(&self, get_badge: Option<&mut Badge>) -> Info;

    fn reply_recv(&self, info: Info, badge: Option<&mut Badge>) -> Info;
}

Syscalls that recieve on a capability

Required methods

fn recv(&self, get_badge: Option<&mut Badge>) -> Info

Recieves message.

Blocks if no sender or notification pending Limited to end-point and ntfn caps Raises fault with other cap types (sel4 manual: 6.2) if get_badge, returns (_, Some(badge_number)) to get info on where it came from

fn nb_recv(&self, get_badge: Option<&mut Badge>) -> Info

As with recv, but does not block.

fn reply_recv(&self, info: Info, badge: Option<&mut Badge>) -> Info

High performance chaining of reply() then recv() with atomic transition between repyl and recv phase

Loading content...