Module sel4_rust_doc::kernel_api::types::err

Structs

AllowedRange

DepthMismatchData

GuardMismatchData

Enums

LookupFailureDesc
Deserialised MessageInfo data following a failed CNode lookup.

SeL4Error
This deserialises the error data contained in a message::Info

Type Definitions

ForSourceCap

InvalidCapPtr
The capability used with a system call in an invalid manner

InvalidNum
An invalid number used in a system call

MemoryAvailable
Actual memory available