Enum sel4_rust_doc::kernel_api::types::err::SeL4Error

pub enum SeL4Error {
    InvalidArgument(InvalidNum),
    InvalidCapability(InvalidCapPtr),
    IllegalOperation,
    RangeError(AllowedRange),
    AlignmentError,
    FailedLookup(ForSourceCap, LookupFailureDesc),
    TruncatedMessage,
    DeleteFirst,
    RevokeFirst,
    NotEnoughMemory(MemoryAvailable),
}

This deserialises the error data contained in a message::Info

Variants

InvalidArgument(InvalidNum)

A non-capability argument is invalid

InvalidCapability(InvalidCapPtr)

A capability is invoked by invalid means

IllegalOperation

A requested operation is not permitted

RangeError(AllowedRange)

An argument is out of the allowed range

AlignmentError

A supplied argument does not meet the allignment requirements

FailedLookup(ForSourceCap, LookupFailureDesc)

A capability could not be looked up

TruncatedMessage

Too few message words or capabilities were sent in the message

DeleteFirst

A destination slot specified in the syscall arguments is occupied

RevokeFirst

The object currently has other object derived from it

The requested invocation can be performed when derived objects are deleted, or revoke is invoked on this capability

NotEnoughMemory(MemoryAvailable)

Insufficient unallocated space to complete a retype request