Module sel4_rust_doc::kernel_api::object_interfaces

Implement OS services with the creation, manipulation and combination of these.

Under the hood, all the methods on these objects are convenience wrappers arround methods documented in the syscalls module. For example, UntypedMemory::retype runs the appropriate configuration before calling syscalls::SeL4Send::send

Modules

capability_space
A threads capability management system.

endpoints
The kernel object that Threads use to send/recieve IPC messages.

notifications
A simple, non-blocking signalling mechanism that logically represents a set of binary semaphores.

reply
Reply is a special, “free” capability that is saved to a single, dedicated slot in a thread after it recieves a message. reply is equivilant to CapSpace::save_caller followed by a SeL4Send::send

thread_control_block
Representing execution context, and an applications processor usage.

untyped_memory
A dynamically sized object from which all other kernel objects are created.

vspace
ASIDPool and ASIDControl for tracking status of an address space.

Enums

Objects
These objects represent the set of service primitives provided by the kernel