Trait sel4_rust_doc::kernel_api::syscalls::SeL4SendRecv

pub trait SeL4SendRecv: SeL4Send + SeL4Recv {
    fn call(&self, info: Info) -> Info;
}

Allows send and recieve in a single call, thus improving performance.

Required methods

fn call(&self, info: Info) -> Info

Combine send and recv with key difference

On caps to non-endpoint kernel-services, can ret an error, or other msg through the reply

When sent msg goes to another thread (via end-point), as well as the normal send, the kernal also deposits a reply cap to a dedicated slot in the recieving TCB

A reply cap is single-use cap allowing for a reply msg to be sent and wake up the caller (and invalidated by the kernel once used.

See also section 4.2.4 of the sel4 manual

Loading content...